// projects/ng-avataaars-library/src/lib/custom-avatar/options/top-type.ts
export enum TopType {
  NoHair = 'NoHair',
  Eyepatch = 'Eyepatch',
  Hat = 'Hat',
  Hijab = 'Hijab',
  Turban = 'Turban',
  WinterHat1 = 'WinterHat1',
  WinterHat2 = 'WinterHat2',
  WinterHat3 = 'WinterHat3',
  WinterHat4 = 'WinterHat4',
  LongHairBigHair = 'LongHairBigHair',
  LongHairBob = 'LongHairBob',
  LongHairBun = 'LongHairBun',
  LongHairCurly = 'LongHairCurly',
  LongHairCurvy = 'LongHairCurvy',
  LongHairDreads = 'LongHairDreads',
  LongHairFrida = 'LongHairFrida',
  LongHairFro = 'LongHairFro',
  LongHairFroBand = 'LongHairFroBand',
  LongHairNotTooLong = 'LongHairNotTooLong',
  LongHairShavedSides = 'LongHairShavedSides',
  LongHairMiaWallace = 'LongHairMiaWallace',
  LongHairStraight = 'LongHairStraight',
  LongHairStraight2 = 'LongHairStraight2',
  LongHairStraightStrand = 'LongHairStraightStrand',
  ShortHairDreads01 = 'ShortHairDreads01',
  ShortHairDreads02 = 'ShortHairDreads02',
  ShortHairFrizzle = 'ShortHairFrizzle',
  ShortHairShaggy = 'ShortHairShaggy',
  ShortHairShaggyMullet = 'ShortHairShaggyMullet',
  ShortHairShortCurly = 'ShortHairShortCurly',
  ShortHairShortFlat = 'ShortHairShortFlat',
  ShortHairShortRound = 'ShortHairShortRound',
  ShortHairShortWaved = 'ShortHairShortWaved',
  ShortHairSides = 'ShortHairSides',
  ShortHairTheCaesar = 'ShortHairTheCaesar',
  ShortHairTheCaesarSidePart = 'ShortHairTheCaesarSidePart'
}