// projects/ng-avataaars-library/src/lib/custom-avatar/options/clothe-type.ts
export enum ClotheType {
  BlazerShirt = 'BlazerShirt',
  BlazerSweater = 'BlazerSweater',
  CollarSweater = 'CollarSweater',
  GraphicShirt = 'GraphicShirt',
  Hoodie = 'Hoodie',
  Overall = 'Overall',
  ShirtCrewNeck = 'ShirtCrewNeck',
  ShirtScoopNeck = 'ShirtScoopNeck',
  ShirtVNeck = 'ShirtVNeck'
}