// projects/ng-avataaars-library/src/lib/custom-avatar/options/eye-type.ts
export enum EyeType {
  Close = 'Close',
  Cry = 'Cry',
  CryMirrored = 'CryMirrored',
  Default = 'Default',
  Dizzy = 'Dizzy',
  EyeRoll = 'EyeRoll',
  Happy = 'Happy',
  Hearts = 'Hearts',
  Side = 'Side',
  SideMirrored = 'SideMirrored',
  Squint = 'Squint',
  Surprised = 'Surprised',
  Wink = 'Wink',
  WinkMirrored = 'WinkMirrored',
  WinkWacky = 'WinkWacky',
  WinkWackyMirrored = 'WinkWackyMirrored'
}